create global application command, commands mod
