+++ src/discord/server/commands/mod.rs
@@ -1,0 +1,2 @@
+pub mod create;
+pub mod utilities;
