--- src/discord/server/event_handlers/mod.rs
@@ -1,2 +1,0 @@
-mod create;
-mod utilities;
