--- src/discord/server/event_handlers/utilities.rs
@@ -1,10 +1,0 @@
-use serenity::model::prelude::interaction::application_command::{
-	CommandDataOption, CommandDataOptionValue,
-};
-
-pub fn _extract_string_option_value(option: &CommandDataOption) -> Option<String> {
-	match option.resolved.as_ref().unwrap() {
-		CommandDataOptionValue::String(value) => Some(value.clone()),
-		_ => None,
-	}
-}
